fix: convert price type to int

+++ src/apps/orders/api/serializers.py
@@ -6,6 +6,6 @@ class OrderSerializer(serializers.Serializer):
     product_name = serializers.CharField(max_length=255)
     quantity = serializers.IntegerField(min_value=1)
     unit_price = serializers.FloatField(write_only=True)
-    total_price = serializers.DecimalField(max_digits=12, decimal_places=2, read_only=True)
+    total_price = serializers.IntegerField(read_only=True)
     created_at = serializers.DateTimeField(read_only=True)
     updated_at = serializers.DateTimeField(read_only=True)
